rename sync to global
