+++ src/global.rs
